requirements.txt
argon2-cffi
discord_webhook
Flask
flask_cors
flask_jwt_extended
imutils
mysql-connector-python
paho_mqtt
pyserial
python-dotenv
markupsafe==2.0.1
